Moved BlobContainerWithFallback to io-sign package
