+++ apps/io-func-sign-issuer/src/infra/azure/storage/upload.ts
@@ -16,8 +16,6 @@ import {
   withPermissions
 } from "@io-sign/io-sign/infra/azure/storage/blob";
 
-import { BaseContainerClientWithFallback } from "@pagopa/azure-storage-migration-kit";
-
 import { GetUploadUrl, UploadUrl } from "../../../upload";
 
 import { FileStorage } from "../../../upload";
@@ -100,63 +98,3 @@ export const makeGetUploadUrl =
       ),
       TE.chainEitherKW(validate(UploadUrl, "Invalid SAS Url generated."))
     );
-
-// FileStorage backed by a container with read-fallback (migration kit).
-// Writes always go to the primary storage account (ITN).
-// Reads (exists, download, getUrl) try primary first, falling back to the
-// secondary storage account (WEU) when the blob is not yet in ITN.
-export class BlobStorageFileStorageWithFallback implements FileStorage {
-  #containerClient: BaseContainerClientWithFallback;
-
-  constructor(containerClient: BaseContainerClientWithFallback) {
-    this.#containerClient = containerClient;
-    this.exists = this.exists.bind(this);
-    this.download = this.download.bind(this);
-    this.createFromUrl = this.createFromUrl.bind(this);
-    this.remove = this.remove.bind(this);
-    this.getUrl = this.getUrl.bind(this);
-  }
-
-  exists(filename: string) {
-    const blobClient = this.#containerClient.getBlobClient(filename);
-    return TE.tryCatch(
-      () => blobClient.exists(),
-      () => new Error("The specified Blob does not exist.")
-    );
-  }
-
-  download(filename: string) {
-    const blobClient = this.#containerClient.getBlobClient(filename);
-    return TE.tryCatch(() => blobClient.downloadToBuffer(), toError);
-  }
-
-  // Writes to primary (ITN) only.
-  createFromUrl(url: string, filename: string) {
-    const blobClient = this.#containerClient.getBlobClient(filename);
-    return copyFromUrl(url)(blobClient.primaryBlobClient);
-  }
-
-  // Deletes from both primary (ITN) and fallback (WEU) if present.
-  remove(filename: string) {
-    const blobClient = this.#containerClient.getBlobClient(filename);
-    return pipe(
-      TE.tryCatch(() => blobClient.deleteIfExists(), toError),
-      TE.map(constVoid)
-    );
-  }
-
-  // Generates a SAS URL from primary (ITN) if the blob exists there,
-  // otherwise falls back to secondary (WEU).
-  getUrl(filename: string) {
-    const blobClient = this.#containerClient.getBlobClient(filename);
-    const options = pipe(
-      defaultBlobGenerateSasUrlOptions(),
-      withPermissions("r"),
-      withExpireInMinutes(5)
-    );
-    return TE.tryCatch(
-      () => blobClient.generateSasUrl(options),
-      () => new Error("Unable to generate the SAS Url for the specified Blob.")
-    );
-  }
-}
